+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix corruption while loading module
 - [#217] - Remove redundant call (thanks [@nyanpasu64])
 - [#216], [#218] - Fix some compilation warings (thanks [@nyanpasu64], [@OPNA2608])
+- [#210] - Finish editing an instrument name when leaving focus (thanks [@OPNA2608])
 - Modify appearance slightly
 
 [#213]: https://github.com/rerrahkr/BambooTracker/issues/213
@@ -20,6 +21,7 @@
 [#217]: https://github.com/rerrahkr/BambooTracker/pull/217
 [#216]: https://github.com/rerrahkr/BambooTracker/issues/216
 [#218]: https://github.com/rerrahkr/BambooTracker/issues/218
+[#210]: https://github.com/rerrahkr/BambooTracker/issues/210
 
 ## v0.4.2 (2020-05-10)
 ### Added